requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-lru==2.0.4
attrs==23.2.0
autopep8==2.1.0
beautifulsoup4==4.12.3
black==25.1.0
boolean.py==5.0
bs4==0.0.2
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
cssselect==1.2.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distro==1.9.0
fake-useragent==1.5.1
filelock==3.18.0
flake8==7.3.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.10
importlib_metadata==7.1.0
license-expression==30.4.4
lxml==5.2.1
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
meta_ai_api==1.1.7
msgpack==1.1.1
multidict==6.0.5
mypy_extensions==1.1.0
oauthlib==3.2.2
openai==0.28.0
packageurl-python==0.17.3
packaging==25.0
parse==1.20.1
pathspec==0.12.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
propcache==0.3.2
py-serializable==2.1.0
pycodestyle==2.14.0
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.1.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pyppeteer==2.0.0
pyquery==2.0.0
python-decouple==3.8
requests==2.32.4
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==14.1.0
ruff==0.12.7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
toml==0.10.2
tqdm==4.67.1
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==1.26.20
w3lib==2.1.2
websockets==10.4
yarl==1.20.1
zipp==3.18.1
